# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для системы управления событиями EventFlow",
        "title": "EventFlow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@eventflow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/event_registrations": {
            "get": {
                "description": "Возвращает список всех регистраций участников на события с пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventRegistrations"
                ],
                "summary": "Получить список регистраций на события",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Пагинация [start, end]",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка [field, order]",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventRegistration"
                            }
                        },
                        "headers": {
                            "Content-Range": {
                                "type": "string",
                                "description": "Диапазон записей"
                            },
                            "X-Total-Count": {
                                "type": "string",
                                "description": "Общее количество записей"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает регистрацию участника на событие и автоматически создает тикет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventRegistrations"
                ],
                "summary": "Зарегистрировать участника на событие",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "registration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEventRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.EventRegistration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/events": {
            "get": {
                "description": "Возвращает список событий с поддержкой пагинации и фильтрации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Получить список событий",
                "parameters": [
                    {
                        "type": "string",
                        "example": "[0, 24]",
                        "description": "Пагинация [start, end]",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "[\"id\", \"ASC\"]",
                        "description": "Сортировка [field, order]",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "{\"category_id\": 1}",
                        "description": "Фильтрация {field: value}",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по категории",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате начала (\u003e= start_date)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате окончания (\u003c= end_date)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по статусу публикации (published/draft)",
                        "name": "publish_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список событий",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        },
                        "headers": {
                            "Content-Range": {
                                "type": "string",
                                "description": "Диапазон записей"
                            },
                            "X-Total-Count": {
                                "type": "string",
                                "description": "Общее количество записей"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новое событие в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Создать новое событие",
                "parameters": [
                    {
                        "description": "Данные события",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданное событие",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateEventRegistrationRequest": {
            "type": "object",
            "required": [
                "event_id",
                "participant_id",
                "status"
            ],
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "participant_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.CreateEventRequest": {
            "type": "object",
            "required": [
                "category_id",
                "end_time",
                "event_type",
                "publish_status",
                "start_time",
                "title"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "publish_status": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.Event": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "publish_status": {
                    "description": "\"draft\" или \"published\"",
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.EventRegistration": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "participant_id": {
                    "type": "integer"
                },
                "registered_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}